requirements.txt
argparse
requests
beautifulsoup4
python-pptx
openai